Prover pass 2: u32 CSR, 16-row h_i scatter, parallel Spartan validate/bind
All three changes are value-exact — the proof digests are byte-stable
at every shape (2^15 c384138e..., 2^18 7eda070c..., 2^20 b4b731f2...),
and the verifier's code paths keep their sequential behavior (parallel
branches engage only above thresholds no verifier input reaches, and
bind/eq parallelism is prover-only).

(1) CSR footprint: F2CellMap stores u32 row offsets (nnz <= u32::MAX
    enforced as ShapeTooLarge; the digest stream still writes offsets
    as u64 LE, so digests are unchanged); the digest hashes through
    bounded staging chunks instead of whole-array transient copies;
    try_from_csr validates with one parallel sweep (sequential
    canonical-first-error re-scan only on failure); cm_and_map fills
    its CSR range-parallel from closed-form entry offsets — identical
    rows, entries, offsets, digest. At 2^20: live heap 3409 -> 2897 MiB
    (-512 = exactly 4 B x 2^27 offsets), peak 4449 -> 4097 MiB. The
    compressor tax on the remaining 1.2 GiB persists on a 16 GB box;
    closed-form structured maps (no stored CSR) stay the next lever.

(2) h_i scatter: 16-row four-Russians blocks (four subset tables, two
    8x8 transposes, ONE accumulator RMW per output bit per 16 rows —
    half the 8-row block's RMWs at the same table-build cost). hs
    12.9 -> 11.7 ms at 2^15, 112 -> 100 ms at 2^18. A dual-accumulator
    ping-pong for the 8-row block measured neutral and is not kept.

(3) Spartan prove side: validate_elements_field sweeps chunk-parallel
    with a config-pointer fast path and validate_assignment's padding
    scan parallelized (sp:validate 33 -> 9.5 ms at 2^18);
    bind_and_batch evaluates columns in parallel over a level-parallel
    prover-side eq table (sp:bind 27.7 -> 15.2 ms); sp:validate /
    sp:bind / sp:inner / sp:eq / sp:outer prof scopes added.
    spartan_prove 13.4 -> 10.0 ms at 2^15. sp:inner is untouched
    (already parallel) and remains the box-pressure-sensitive phase.

2^15 idle probes (settled, two runs): prove 61.3 -> 58.6 ms
(spartan 10.0, hs 11.6-11.9, a' 9.2-9.4, apply 1.4). 2^18: prove
~402 ms clean. 2^20 bench medians: prove 2608 ms, verify 735 ms, setup
relation+map ~2.3 s (box-noise band), heap live 2897 / peak 4097 MiB.
Suite: 196 tests green, checked + unchecked.

+++ tests/virtual_open.rs
@@ -366,7 +366,8 @@ fn virtual_open_identity_fast_path() {
         lig_configs(packed_vars(&p), LigConfig::Adhoc { log_batch: 2, log_inv_rate: 2 }).unwrap();
     let rows_f = f_rows(&p, 0x1D_FA57);
     let n = cell_count(&p);
-    let map = F2CellMap::try_from_csr(n, n, (0..=n).collect(), (0..n as u32).collect()).unwrap();
+    let map =
+        F2CellMap::try_from_csr(n, n, (0..=n as u32).collect(), (0..n as u32).collect()).unwrap();
     assert!(map.is_identity());
     let hint = commit_rs_ligerito_rows(&p, rows_f.clone(), &pc);
